# docs/api-coverage/openapi-non-prod.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
...
  "paths" : {
...
    "/api/agent/runs/{runUuid}" : {
      "get" : {
        "operationId" : "getRun_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "runUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAgentRunVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "agent-controller" ]
      }
    },
...
    "/api/v5/projects/{projectUuid}/generation-runs" : {
      "post" : {
        "operationId" : "create",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateGenerationRunRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGenerationRun"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "generation-run-controller" ]
      }
    },
    "/api/v5/projects/{projectUuid}/generation-runs/{runUuid}" : {
      "get" : {
        "operationId" : "get",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "runUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGenerationRun"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "generation-run-controller" ]
      }
    },
    "/api/v5/projects/{projectUuid}/generation-runs/{runUuid}/approval" : {
      "post" : {
        "operationId" : "approve",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "runUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GenerationApprovalRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGenerationApprovalOutcome"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "generation-run-controller" ]
      }
    },
    "/api/v5/projects/{projectUuid}/generation-runs/{runUuid}/artifact" : {
      "get" : {
        "operationId" : "artifact",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "runUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "byte",
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "generation-run-controller" ]
      }
    },
    "/api/v5/projects/{projectUuid}/generation-runs/{runUuid}/build" : {
      "post" : {
        "operationId" : "build",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "runUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "expectedVersion",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGenerationBuildOutcome"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "generation-run-controller" ]
      }
    },
    "/api/v5/projects/{projectUuid}/generation-runs/{runUuid}/preview-artifact" : {
      "get" : {
        "operationId" : "previewArtifact",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "runUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "byte",
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "generation-run-controller" ]
      }
    },
...
    "/api/v5/projects/{projectUuid}/generation-runs/{runUuid}/prototype-version-compatibility" : {
      "get" : {
        "operationId" : "inspectPrototypeCompatibility",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "runUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGenerationPrototypeBridgeResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "generation-run-controller" ]
      }
    },
    "/api/v5/projects/{projectUuid}/generation-runs/{runUuid}/release" : {
      "post" : {
        "operationId" : "release",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "runUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "expectedVersion",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGenerationRun"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "generation-run-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AgentRunVO" : {
        "properties" : {
          "agentType" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "errorCategory" : {
            "type" : "string"
          },
          "errorMessage" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "mockState" : {
            "type" : "string"
          },
          "modelName" : {
            "type" : "string"
          },
          "projectId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "projectUuid" : {
            "type" : "string"
          },
          "provider" : {
            "type" : "string"
          },
          "runUuid" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "timeTakenMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "traceId" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseAgentRunVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/AgentRunVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseGenerationApprovalOutcome" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/GenerationApprovalOutcome"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponseGenerationBuildOutcome" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/GenerationBuildOutcome"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponseGenerationPrototypeBridgeResponse" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/GenerationPrototypeBridgeResponse"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponseGenerationRun" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/GenerationRun"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CreateGenerationRunRequest" : {
        "properties" : {
          "spec" : {
            "$ref" : "#/components/schemas/JsonNode"
          }
        },
        "required" : [ "spec" ],
        "type" : "object"
      },
...
      "GenerationApprovalOutcome" : {
        "properties" : {
          "actorUserId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "approvalUuid" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "decision" : {
            "type" : "string"
          },
          "generationRunUuid" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "reused" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "GenerationApprovalRequest" : {
        "properties" : {
          "decision" : {
            "minLength" : 1,
            "pattern" : "APPROVED|REJECTED",
            "type" : "string"
          },
          "reason" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "decision", "reason" ],
        "type" : "object"
      },
      "GenerationBuildOutcome" : {
        "properties" : {
          "buildLogDigest" : {
            "type" : "string"
          },
          "cocosExitCode" : {
            "format" : "int32",
            "type" : "integer"
          },
          "cocosOutputDigest" : {
            "type" : "string"
          },
          "generationRunUuid" : {
            "type" : "string"
          },
          "packageDigest" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "GenerationPrototypeBridgeResponse" : {
        "properties" : {
          "compatible" : {
            "type" : "boolean"
          },
          "prototypeVersionUuid" : {
            "type" : "string"
          },
          "reasons" : {
            "items" : {
              "$ref" : "#/components/schemas/Incompatibility"
            },
            "type" : "array"
          },
          "reused" : {
            "type" : "boolean"
          },
          "source" : {
            "$ref" : "#/components/schemas/SourceSummary"
          }
        },
        "type" : "object"
      },
      "GenerationRun" : {
        "properties" : {
          "buildAttempt" : {
            "format" : "int32",
            "type" : "integer"
          },
          "buildClaimExpiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "buildRequestJson" : {
            "type" : "string"
          },
          "canonicalSpecJson" : {
            "type" : "string"
          },
          "completedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "diagnosticsJson" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "idempotencyKey" : {
            "type" : "string"
          },
          "packageDigest" : {
            "type" : "string"
          },
          "projectId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "requestFingerprint" : {
            "type" : "string"
          },
          "runUuid" : {
            "type" : "string"
          },
          "runtimeIrDigest" : {
            "type" : "string"
          },
          "runtimeIrJson" : {
            "type" : "string"
          },
          "sourceDigest" : {
            "type" : "string"
          },
          "stateVersion" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "Incompatibility" : {
        "properties" : {
          "actual" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "expected" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "JsonNode" : { },
...
      "SourceSummary" : {
        "properties" : {
          "runUuid" : {
            "type" : "string"
          },
          "runtimeIrDigest" : {
            "type" : "string"
          },
          "sourceDigest" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  }
}